split EventArgs into individual files
