+++ src/Soulseek/Common/EventArgs/BrowseEventArgs.cs
@@ -31,43 +31,4 @@ namespace Soulseek
         /// </summary>
         public string Username { get; }
     }
-
-    /// <summary>
-    ///     Event arguments for events raised by receipt of browse response data.
-    /// </summary>
-    public class BrowseProgressUpdatedEventArgs : BrowseEventArgs
-    {
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="BrowseProgressUpdatedEventArgs"/> class.
-        /// </summary>
-        /// <param name="username">The username associated with the event.</param>
-        /// <param name="bytesTransferred">The total number of bytes transfereed.</param>
-        /// <param name="size">The total expected length of the data transfer.</param>
-        public BrowseProgressUpdatedEventArgs(string username, long bytesTransferred, long size)
-            : base(username)
-        {
-            BytesTransferred = bytesTransferred;
-            Size = size;
-        }
-
-        /// <summary>
-        ///     Gets the total number of bytes transferred.
-        /// </summary>
-        public long BytesTransferred { get; }
-
-        /// <summary>
-        ///     Gets the number of remaining bytes to be transferred.
-        /// </summary>
-        public long BytesRemaining => Size - BytesTransferred;
-
-        /// <summary>
-        ///     Gets the progress of the data transfer as a percentage of current and total data length.
-        /// </summary>
-        public double PercentComplete => (BytesTransferred / (double)Size) * 100d;
-
-        /// <summary>
-        ///     Gets the total expected length of the data transfer.
-        /// </summary>
-        public long Size { get; }
-    }
 }
